requirements.txt
altair==4.2.0
altgraph==0.17.2
ansi2html==1.8.0
anyio==3.6.1
app==0.0.1
appdirs==1.4.4
apps==0.0.0.0.dev20200918
apriori==1.0.0
apyori==1.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.5
branca==0.5.0
Brotli==1.0.9
bs4==0.0.1
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
dash==2.6.0
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-extensions==0.1.5
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.5
docopt==0.6.2
EditorConfig==0.12.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
facebook-scraper==0.2.57
fake-useragent==0.1.11
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Compress==1.12
folium==0.12.1.post1
fonttools==4.30.0
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.52
geopandas==0.11.1
geopy==2.2.0
geotext==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gspread==5.4.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
insta-scrape==2.1.2
instagram-scraper==1.11.0
instaloader==4.9.2
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
jsbeautifier==1.14.4
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-dash==0.4.2
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
jupyterthemes==0.20.0
kiwisolver==1.3.2
lesscpy==0.15.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlxtend==0.19.0
more-itertools==8.13.0
MultiApp==1.0
munch==2.5.0
nbclassic==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
patsy==0.5.2
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.0.1
pipwin==0.5.2
plotly==5.9.0
plotly-express==0.4.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pydeck==0.7.1
pyee==8.2.2
Pygments==2.12.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyjsparser==2.7.1
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyppeteer==1.0.2
PyPrind==2.11.3
pyproj==3.3.1
pyquery==1.4.3
pyrsistent==0.18.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
rake-nltk==1.0.6
regex==2022.3.2
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
retrying==1.3.3
rich==12.5.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
scraper==0.1.0
seaborn==0.11.2
selenium==4.3.0
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
squarify==0.4.3
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.12.0
tenacity==8.0.1
terminado==0.15.0
textblob==0.17.1
